add lambda.js for lambda-izing
