+++ package.json
@@ -13,6 +13,7 @@
     "node": ">=0.14"
   },
   "dependencies": {
+    "@vendia/serverless-express": "^4.0.0-rc.2",
     "autoprefixer": "^10.0.2",
     "express": "^4.17.1",
     "express-validator": "^6.9.2",
