--- test.css
@@ -1,8 +1,0 @@
-:root {
-            --pc-color-1: rgba(0,1,2,.3);
-            --pc-color-2: red;
-        }
-        .foo {
-            color: var(--pc-color-1);
-            background-color: var(--pc-color-2);
-        }
